chore(package): clean up crate a bit
