+++ package/src/package/manifest.rs
@@ -1,0 +1,44 @@
+use alloc::{collections::BTreeSet, vec::Vec};
+use core::fmt;
+
+use miden_assembly::ast::QualifiedProcedureName;
+use miden_core::{Word, utils::DisplayHex};
+
+use crate::Dependency;
+
+// PACKAGE MANIFEST
+// ================================================================================================
+
+/// The manifest of a package, containing the set of package dependencies (libraries or packages)
+/// and exported procedures and their signatures, if known.
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
+#[cfg_attr(test, derive(proptest_derive::Arbitrary))]
+pub struct PackageManifest {
+    /// The set of exports in this package.
+    pub exports: BTreeSet<PackageExport>,
+    /// The libraries (packages) linked against by this package, which must be provided when
+    /// executing the program.
+    pub dependencies: Vec<Dependency>,
+}
+
+/// A procedure exported by a package, along with its digest and signature (will be added after
+/// MASM type attributes are implemented).
+#[derive(Clone, PartialEq, Eq, Ord, PartialOrd)]
+#[cfg_attr(test, derive(proptest_derive::Arbitrary))]
+pub struct PackageExport {
+    /// The fully-qualified name of the procedure exported by this package
+    pub name: QualifiedProcedureName,
+    /// The digest of the procedure exported by this package
+    #[cfg_attr(test, proptest(value = "Word::default()"))]
+    pub digest: Word,
+}
+
+impl fmt::Debug for PackageExport {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let Self { name, digest } = self;
+        f.debug_struct("PackageExport")
+            .field("name", &format_args!("{name}"))
+            .field("digest", &format_args!("{}", DisplayHex::new(&digest.as_bytes())))
+            .finish()
+    }
+}
